// apps/console/src/api/generated/models/problemDetails.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Taskome API
 * Taskome control-plane API
 * OpenAPI spec version: 1.0.0
 */
import type { ProblemDetailsErrorsItem } from "./problemDetailsErrorsItem";

export interface ProblemDetails {
  code: string;
  detail: string;
  errors?: ProblemDetailsErrorsItem[];
  instance: string;
  requestId: string;
  status: number;
  title: string;
  type: string;
}